+++ src/main/java/com/skycat/mystical/mixin/SheepEntityMixin.java
@@ -15,15 +15,18 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
 @Mixin(SheepEntity.class)
 public abstract class SheepEntityMixin implements EntityLike {
-    @Shadow public abstract void setColor(DyeColor color);
-
     @Inject(method = "onEatingGrass", at = @At("TAIL"))
     public void afterEatGrass(CallbackInfo ci) {
         // https://www.reddit.com/r/fabricmc/comments/kfrsq7/comment/ggactl1/
-        if (!(Mystical.isClientWorld() && Mystical.getHavenManager().isInHaven(getBlockPos())) &&
-                (Mystical.isClientWorld() && Mystical.getSpellHandler().isConsequenceActive(SheepColorChangeConsequence.class) && Utils.percentChance(Mystical.CONFIG.sheepColorChange.chance()))) {
+        if (!Mystical.isClientWorld() &&
+                !Mystical.getHavenManager().isInHaven(getBlockPos()) &&
+                Mystical.getSpellHandler().isConsequenceActive(SheepColorChangeConsequence.class) &&
+                Utils.percentChance(Mystical.CONFIG.sheepColorChange.chance())) {
             setColor(Util.getRandom(DyeColor.values(), Mystical.MC_RANDOM));
             Utils.log(Utils.translateString("text.mystical.consequence.sheepColorChange.fired"), Mystical.CONFIG.sheepColorChange.logLevel());
         }
     }
+
+    @Shadow
+    public abstract void setColor(DyeColor color);
 }