+++ src/main/java/com/skycat/mystical/mixin/SkulkShriekerBlockEntityMixin.java
@@ -13,7 +13,9 @@ public abstract class SkulkShriekerBlockEntityMixin extends BlockEntityMixin {
 
     @ModifyConstant(method = "trySpawnWarden", constant = @Constant(intValue = 4))
     private int warningsBeforeSpawning(int constant) { // TODO: This currently tests if the shrieker is in a haven. Maybe this should be the player instead (set player strikes to 4 whenever they gain 1?)
-        if ((Mystical.isClientWorld() && Mystical.getSpellHandler().isConsequenceActive(OneStrikeWardensConsequence.class) && !(Mystical.isClientWorld() && Mystical.getHavenManager().isInHaven(getPos())))) {
+        if (!Mystical.isClientWorld() &&
+                Mystical.getSpellHandler().isConsequenceActive(OneStrikeWardensConsequence.class) &&
+                !Mystical.getHavenManager().isInHaven(getPos())) {
             Utils.log(Utils.translateString(OneStrikeWardensConsequence.FACTORY.getDescriptionKey()), Mystical.CONFIG.oneStrikeWardens.logLevel());
             return 1; // TODO: Logging
         }