+++ src/main/java/com/skycat/mystical/mixin/SpawnHelperMixin.java
@@ -31,10 +31,31 @@ import org.spongepowered.asm.mixin.injection.callback.LocalCapture;
 @Mixin(SpawnHelper.class)
 public abstract class SpawnHelperMixin {
 
+    @Inject(method = "spawnEntitiesInChunk(Lnet/minecraft/entity/SpawnGroup;Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/world/chunk/Chunk;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/world/SpawnHelper$Checker;Lnet/minecraft/world/SpawnHelper$Runner;)V",
+            at = @At(value = "INVOKE", target = "Lnet/minecraft/server/world/ServerWorld;spawnEntityAndPassengers(Lnet/minecraft/entity/Entity;)V", shift = At.Shift.AFTER),
+            locals = LocalCapture.CAPTURE_FAILSOFT)
+    // Tried using ModifyVariable, but locations weren't set up yet. // TY llamalad7, benonardo // This could be better as a WrapOperation
+    private static void giveSpeedBoost(SpawnGroup group, ServerWorld world, Chunk chunk, BlockPos pos, SpawnHelper.Checker checker, SpawnHelper.Runner runner, CallbackInfo ci, StructureAccessor structureAccessor, ChunkGenerator chunkGenerator, int i, BlockState blockState, BlockPos.Mutable mutable, int j, int k, int l, int m, int n, SpawnSettings.SpawnEntry spawnEntry, EntityData entityData, int o, int p, int q, double d, double e, PlayerEntity playerEntity, double f, MobEntity mobEntity) {
+        // Most of the parameters are unused :(
+        for (Spell spell : Mystical.getSpellHandler().spellsOfConsequenceType(TurboMobsConsequence.class)) { // Inside this, there must be an active TurboMobsConsequence spell
+            if (!Mystical.isClientWorld() &&
+                    Mystical.getHavenManager().isInHaven(mobEntity) &&
+                    ((TurboMobsConsequence) spell.getConsequence()).entityType.equals(mobEntity.getType()) { // Make sure it applies to this type // TODO: Chance
+                EntityAttributeInstance attributeInstance = mobEntity.getAttributeInstance(EntityAttributes.GENERIC_MOVEMENT_SPEED);
+                if (attributeInstance != null) {
+                    attributeInstance.addPersistentModifier(new EntityAttributeModifier("Mystical speed boost", 0.5, EntityAttributeModifier.Operation.MULTIPLY_BASE)); // TODO: Config
+                    Utils.log(Utils.translateString(TurboMobsConsequence.FACTORY.getDescriptionKey()), Mystical.CONFIG.turboMobs.logLevel());
+                }
+            }
+        }
+    }
+
     @ModifyArg(method = "spawnEntitiesInChunk(Lnet/minecraft/entity/SpawnGroup;Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/world/chunk/Chunk;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/world/SpawnHelper$Checker;Lnet/minecraft/world/SpawnHelper$Runner;)V",
             at = @At(value = "INVOKE", target = "Lnet/minecraft/world/SpawnHelper;createMob(Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/entity/EntityType;)Lnet/minecraft/entity/mob/MobEntity;"), index = 1)
     private static EntityType swapMobs(EntityType entityType) {
-        if (!(Mystical.isClientWorld() && Mystical.getSpellHandler().isConsequenceActive(MobSpawnSwapConsequence.class))) return entityType;
+        if (!Mystical.isClientWorld() && Mystical.getSpellHandler().isConsequenceActive(MobSpawnSwapConsequence.class)) {
+            return entityType;
+        }
         if (entityType == EntityType.SKELETON) return EntityType.WITHER_SKELETON; // Skeleton <-> Wither skeleton
         if (entityType == EntityType.WITHER_SKELETON) return EntityType.SKELETON;
         if (entityType == EntityType.ZOMBIE) return EntityType.ZOMBIFIED_PIGLIN; // Zombie <-> Zombified piglin
@@ -45,22 +66,5 @@ public abstract class SpawnHelperMixin {
         if (entityType == EntityType.BLAZE) return EntityType.CREEPER;
 
         return entityType;
-    }
-
-    @Inject(method = "spawnEntitiesInChunk(Lnet/minecraft/entity/SpawnGroup;Lnet/minecraft/server/world/ServerWorld;Lnet/minecraft/world/chunk/Chunk;Lnet/minecraft/util/math/BlockPos;Lnet/minecraft/world/SpawnHelper$Checker;Lnet/minecraft/world/SpawnHelper$Runner;)V",
-            at = @At(value = "INVOKE", target = "Lnet/minecraft/server/world/ServerWorld;spawnEntityAndPassengers(Lnet/minecraft/entity/Entity;)V", shift = At.Shift.AFTER),
-            locals = LocalCapture.CAPTURE_FAILSOFT) // Tried using ModifyVariable, but locations weren't set up yet. // TY llamalad7, benonardo // This could be better as a WrapOperation, but that's yet another dependency :/
-    private static void giveSpeedBoost(SpawnGroup group, ServerWorld world, Chunk chunk, BlockPos pos, SpawnHelper.Checker checker, SpawnHelper.Runner runner, CallbackInfo ci, StructureAccessor structureAccessor, ChunkGenerator chunkGenerator, int i, BlockState blockState, BlockPos.Mutable mutable, int j, int k, int l, int m, int n, SpawnSettings.SpawnEntry spawnEntry, EntityData entityData, int o, int p, int q, double d, double e, PlayerEntity playerEntity, double f, MobEntity mobEntity) {
-        // Most of the parameters are unused :(
-        for (Spell spell : Mystical.getSpellHandler().spellsOfConsequenceType(TurboMobsConsequence.class)) { // Inside this, there must be an active TurboMobsConsequence spell
-            if (!(Mystical.isClientWorld() && Mystical.getHavenManager().isInHaven(mobEntity)) && // Ensure outside haven
-                    ((TurboMobsConsequence) spell.getConsequence()).entityType.equals(mobEntity.getType())) { // Make sure it applies to this type // TODO: Chance
-                EntityAttributeInstance attributeInstance = mobEntity.getAttributeInstance(EntityAttributes.GENERIC_MOVEMENT_SPEED);
-                if (attributeInstance != null) {
-                    attributeInstance.addPersistentModifier(new EntityAttributeModifier("Mystical speed boost", 0.5, EntityAttributeModifier.Operation.MULTIPLY_BASE)); // TODO: Config
-                    Utils.log(Utils.translateString(TurboMobsConsequence.FACTORY.getDescriptionKey()), Mystical.CONFIG.turboMobs.logLevel());
-                }
-            }
-        }
     }
 }