+++ src/main/java/com/skycat/mystical/mixin/SummonVexGoalMixin.java
@@ -28,15 +28,12 @@ public abstract class SummonVexGoalMixin {
         return pos;
     }
 
-    @WrapOperation(
-            method = "castSpell",
-            at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/EntityType;create(Lnet/minecraft/world/World;)Lnet/minecraft/entity/Entity;")
-    )
+    @WrapOperation(method = "castSpell", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/EntityType;create(Lnet/minecraft/world/World;)Lnet/minecraft/entity/Entity;"))
     public Entity overrideVexSpawning(EntityType instance, World world, Operation<VexEntity> original) { // TODO: Make things not spawn in walls
-        if ((Mystical.isClientWorld() && Mystical.getHavenManager().isInHaven(spawningPos)) ||
-                world.isClient() ||
+        if (Mystical.isClientWorld() || // world.isClient() ||
+                Mystical.getHavenManager().isInHaven(spawningPos) ||
                 spawningPos == null ||
-                !(Mystical.isClientWorld() && (Mystical.isClientWorld() && Mystical.getSpellHandler().isConsequenceActive(RandomEvokerSummonsConsequence.class))) ||
+                !Mystical.getSpellHandler().isConsequenceActive(RandomEvokerSummonsConsequence.class) ||
                 !Utils.percentChance(Mystical.CONFIG.randomEvokerSummons.chance())) {
             return original.call(instance, world);
         }